requirements.txt
fastapi
uvicorn
jinja2
python-multipart
faster-whisper
python-dotenv
